backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# AI/ML
sentence-transformers>=2.6.0
google-generativeai==0.3.2
openai==1.12.0
pinecone-client==3.0.0

# Redis
redis==5.0.1

# Excel/PDF Export
openpyxl==3.1.2
reportlab==4.0.9

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.2.0
